requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
antiword==0.1.0
anyio==4.13.0
argcomplete==1.10.3
attrs==26.1.0
beautifulsoup4==4.8.2
blis==1.3.3
catalogue==2.0.10
certifi==2026.2.25
chardet==3.0.4
charset-normalizer==3.4.7
click==8.3.2
cloudpathlib==0.23.0
compressed-rtf==1.0.7
confection==1.3.3
cuda-bindings==13.2.0
cuda-pathfinder==1.5.1
cuda-toolkit==13.0.2
cymem==2.0.13
dataclasses-json==0.6.7
docx2txt==0.9
ebcdic==2.0.1
extract-msg==0.28.7
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.3.0
greenlet==3.3.2
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
huggingface_hub==1.9.0
idna==3.11
IMAPClient==2.1.0
Jinja2==3.1.6
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
langchain-classic==1.0.3
langchain-community==0.4.1
langchain-core==1.2.26
langchain-experimental==0.4.1
langchain-huggingface==1.2.1
langchain-text-splitters==1.1.1
langsmith==0.7.25
lxml==6.0.2
markdown-it-py==4.0.0
MarkupSafe==3.0.3
marshmallow==3.26.2
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.1
murmurhash==1.0.15
mypy_extensions==1.1.0
networkx==3.6.1
numpy==2.4.4
nvidia-cublas==13.1.0.3
nvidia-cuda-cupti==13.0.85
nvidia-cuda-nvrtc==13.0.88
nvidia-cuda-runtime==13.0.96
nvidia-cudnn-cu13==9.19.0.56
nvidia-cufft==12.0.0.61
nvidia-cufile==1.15.1.6
nvidia-curand==10.4.0.35
nvidia-cusolver==12.0.4.66
nvidia-cusparse==12.6.3.3
nvidia-cusparselt-cu13==0.8.0
nvidia-nccl-cu13==2.28.9
nvidia-nvjitlink==13.0.88
nvidia-nvshmem-cu13==3.4.5
nvidia-nvtx==13.0.85
olefile==0.47
orjson==3.11.8
packaging==26.0
pdfminer.six==20191110
pillow==12.2.0
preshed==3.0.13
propcache==0.4.1
pt_core_news_sm @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_sm-3.8.0/pt_core_news_sm-3.8.0-py3-none-any.whl#sha256=c304fa04db3af73cd08a250feacf560506e15a2ec2469bd1b09f06847f6b455c
pycryptodome==3.23.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.20.0
python-docx==1.2.0
python-dotenv==1.2.2
python-pptx==0.6.23
PyYAML==6.0.3
regex==2026.4.4
requests==2.33.1
requests-toolbelt==1.0.0
rich==14.3.3
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setuptools==82.0.1
shellingham==1.5.4
six==1.12.0
smart_open==7.5.1
sortedcontainers==2.4.0
soupsieve==2.8.3
spacy==3.8.14
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.8.1
SQLAlchemy==2.0.49
srsly==2.5.3
sympy==1.14.0
tenacity==9.1.4
textract==1.6.5
thinc==8.3.13
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.11.0
tqdm==4.67.3
transformers==5.5.0
triton==3.6.0
typer==0.24.1
typing-inspect==0.9.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzlocal==5.3.1
urllib3==2.6.3
uuid_utils==0.14.1
wasabi==1.1.3
weasel==1.0.0
wrapt==2.1.2
xlrd==1.2.0
xlsxwriter==3.2.9
xxhash==3.6.0
yarl==1.23.0
zstandard==0.25.0
